% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots points where accidents for given state.num}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{state number to plot accidents for}

\item{year}{year of data to plot accidents that occured}
}
\value{
a data.frame of filtered data, if there are no rows then returns invisible(NULL)
}
\description{
\code{fars_map_state} plots the state given as an integer
argument and plots accidents for a given year
}
\details{
The function will take a year and state and attempt to plot accidents for a given state number.
If there is no data for a state, the function will stop and throw an error. If there are no accidents
a message will be returning indicating this.
}
\note{
This function depends on map and points functions from the maps and graphics package respectively.
}
\examples{
\dontrun{
  fars_map_state(1, 2013)
}


}
